% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_to_know_me.R
\name{get_to_know_me}
\alias{get_to_know_me}
\title{Characteristic traits}
\usage{
get_to_know_me(name, age, color)
}
\arguments{
\item{name}{a character containing the input name}

\item{age}{a number containing the user's age}

\item{color}{a character containing  the user's favorite color}
}
\value{

}
\description{
Contains one function, get_to_know_me, that prints a statement about the user
}
