% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inches_to_cm.R
\name{inches_to_cm}
\alias{inches_to_cm}
\title{Converts a measurement from inches to centimeters}
\usage{
inches_to_cm(inches, object)
}
\arguments{
\item{inches}{a number representing a measurement in inches}

\item{object}{a string naming the object being measured}
}
\value{

}
\description{
Converts a measurement from inches to centimeters
}
