% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restaurant_location.R
\name{restaurant_location}
\alias{restaurant_location}
\title{Restaurant Location}
\usage{
restaurant_location(restaurant, address)
}
\arguments{
\item{restaurant}{a string indicating the name of a restaurant}

\item{address}{a string containing the address of the restaurant}
}
\value{

}
\description{
Restaurant Location
}
