% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bluepink.R
\name{theme_bluepink}
\alias{theme_bluepink}
\title{A blue and pink theme}
\usage{
theme_bluepink()
}
\value{

}
\description{
This is a theme that includes a light blue and pink color scheme.
}
\examples{

}
